Update I_Test_Driver_FdV.yaml
The current value for RootDocumentId = 'uniqueID^^^^urn:gematik:iti:xds:2023:rootDocumentUniqueId' which maps to "uniqueID + ihe-coding" , doesn't work to set a new assignment for a deny policy. This is not supported by SDK, since SDK is doing the ihe-coding internally for FdV to avoid unnecessary errors. Hence may I suggest on RISE behalf to update the "rootDocumentId" value as only the "UniqueID" or the "referenceIdList"